Governance page
This is mostly finalized, still need to implement the sort and some prep for the API integration, but had a lot of pending changes and wanted to push before heading out.

Also included is a lot of clean up with usage of single or double quotes. OCD but looks good